Moved messaging modal to common components.
- Moved rx/rx-modal partial to modules/m-modal and updated paths.
- Updated modal class names to use va-modal prefix
- Pared rx-modal partials styles down to Rx-specific CSS.
- Added messaging-specific SCSS partials.
- Refactoring modal CSS.
- Adding messaging-modal specific CSS.
- Added modal close button, made it hide-able.

Unblocks department-of-veterans-affairs/vets-website#3091.
