+++ src/js/common/components/Modal.jsx
@@ -12,25 +12,27 @@ class Modal extends React.Component {
   }
 
   render() {
-    const modalCss = classNames('rx-modal', this.props.cssClass);
+    const modalCss = classNames('va-modal', this.props.cssClass);
 
     if (!this.props.visible) {
       return <div/>;
     }
 
+    const closeButton = this.props.hideCloseButton ? '' : (
+      <button
+          className="va-modal-close"
+          type="button"
+          onClick={this.props.onClose}>
+        <i className="fa fa-close"></i>
+        <span className="usa-sr-only">Close this modal</span>
+      </button>
+    );
+
     return (
-      <div
-          className={modalCss}
-          id={this.props.id}>
-        <div className="rx-modal-inner">
-          <button
-              className="rx-modal-close"
-              onClick={this.handleClose}
-              type="button">
-            <i className="fa fa-close"></i>
-            <span className="usa-sr-only">Close</span>
-          </button>
-          <div className="rx-modal-body">
+      <div className={modalCss} id={this.props.id}>
+        <div className="va-modal-inner">
+          {closeButton}
+          <div className="va-modal-body">
             {this.props.contents}
           </div>
         </div>
@@ -44,7 +46,8 @@ Modal.propTypes = {
   cssClass: React.PropTypes.string,
   id: React.PropTypes.string,
   onClose: React.PropTypes.func.isRequired,
-  visible: React.PropTypes.bool.isRequired
+  visible: React.PropTypes.bool.isRequired,
+  hideCloseButton: React.PropTypes.bool
 };
 
 export default Modal;
