+++ src/sass/messaging/messaging.scss
@@ -1,4 +1,5 @@
 @import "../style";
+@import "../modules/m-modal";
 
 $messaging-label-width: 10rem;
 
@@ -8,7 +9,5 @@ $messaging-label-width: 10rem;
 @import "partials/messaging-btn-delete";
 @import "partials/messaging-folder";
 @import "partials/messaging-thread";
+@import "partials/messaging-modal";
 
-// TODO: Swap with common modal styles once
-// they exist.
-@import "../rx/_partials/rx-modal";
