+++ src/sass/messaging/partials/_messaging-modal.scss
@@ -1,0 +1,5 @@
+.messaging-modal {
+  &-title {
+    color: $color-primary; 
+  }
+}
