+++ src/sass/rx/rx.scss
@@ -3,6 +3,7 @@
 // elements version of these files.
 
 @import "../style";
+@import "../modules/m-modal";
 
 // TODO: Contributed these back to core / Vanguard.
 @import "_partials/va-tabs";
